pyproject.toml
[tool.poetry]
name = "automate"
version = "0.0.1"
description = ""
authors = ["Christoph Gerum <christoph.gerum@uni-tuebingen.de>"]

[tool.poetry.dependencies]
python = "^3.6"
pydantic = "^1.2"
"ruamel.yaml" = "^0.16.5"
coloredlogs = "^10.0"
tabulate = "^0.8.6"
invoke = "^1.3"
fabric = "^2.5"
patchwork = "^1.0"
pre-commit = "^1.20"
prompt_toolkit = "^3.0"
mkdocs-material = "^4.5"

[tool.poetry.dev-dependencies]
pur = "^5.2"
mypy = "==0.740.0"
typeshed = "^0.0.1"
pre-commit = "^1.20"
pytest = "^5.3"
pytest-cov = "^2.8"
pytest-monkeytype = "^1.0"
black = {version = "^18.3-alpha.0", allows-prereleases = true}
mkdocs = "^1.0"
pygments = "^2.5"
coverage = {version = "^4.5", extras = ["toml"]}
mock-ssh-server = "^0.8.0"

[tool.poetry.scripts]
automate = 'automate.main:program.run'


[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"

